Use 16bits for match length

+++ src/match_finder.rs
@@ -2,16 +2,14 @@ use std::time::Instant;
 
 use crate::mdma::MdmaIndex;
 
-// TODO: Use a range based MatchGen
 pub fn generate(mdma_index: &MdmaIndex, matches: &mut Vec<Match>) {
     let lcp_array = build_lcp_array(&mdma_index.sa, &mdma_index.buf);
     let timer = Instant::now();
     let mut stack: Vec<MatchGen> = Vec::with_capacity(256);
 
     for index in 0..lcp_array.len() {
-        // TODO: Use u16 for the lz matches
         let lcp = lcp_array[index];
-        let lcp = if lcp > u8::MAX as i32 { u8::MAX } else { lcp as u8 };
+        let lcp = if lcp > u16::MAX as i32 { u16::MAX } else { lcp as u16 };
 
         // Push new matches
         if lcp > stack.last().map_or(1, |m| m.len) {
@@ -72,30 +70,18 @@ pub fn build_lcp_array(sa: &Vec<i32>, buf: &Vec<u8>) -> Vec<i32> {
     return lcp;
 }
 
-// TODO: Maybe find a way to align this?
-// [ ] Check how big the stack can get
-// [ ] Check hwo big the stack can get with len < 256
-// [ ] See if we can group multiple matches together?
-// Elaborating on that last one: When we have matches with the same count and same sa_index, but different lenghts,
-// maybe we don't need to rank all of them (just the longest one?)
-// Even if we can't get around ranking all of them, perhaps we can at least store them in stack more efficiently?
-// ---------------------------
 // MatchGen is a more lightweight struct that only holds the len and sa_index
 // It's used to  minimize the allocations of the match_finder
 // Since we always compute the sa_count field, we can not store it and save a couple of bytes
-// NOTE: range of len is [2-], but range of u8 is [0-255], so we *could* offset the len by 2
-// while storing in match to utilize a bigger range [2-257] if we needed to
-// (but I think the bigger range can't compensate for the extra complexity per access)
-// - same with sa_count
 
 // #[repr(packed(1))]
 struct MatchGen {
     sa_index: u32,
-    len: u8
+    len: u16
 }
 
 impl MatchGen {
-    fn new(sa_index: usize, len: u8) -> Self { Self { sa_index: sa_index as u32, len } }
+    fn new(sa_index: usize, len: u16) -> Self { Self { sa_index: sa_index as u32, len } }
 }
 
 #[derive(Clone)]
@@ -103,7 +89,7 @@ pub struct Match {
     pub self_ref: bool,
     pub sa_index: u32,
     pub sa_count: u32,
-    pub len:      u8
+    pub len:      u16
 }
 
 impl Match {
@@ -111,7 +97,7 @@ impl Match {
         Self { self_ref: true, sa_index: mg.sa_index, sa_count: index as u32 - mg.sa_index + 1, len: mg.len }
     }
 
-    fn with_len(m: &Match, len: u8) -> Self {
+    fn with_len(m: &Match, len: u16) -> Self {
         let mut clone = m.clone();
         clone.len = len;
         return clone;
@@ -133,7 +119,7 @@ pub fn _static_analyze(mdma_index: &MdmaIndex) {
     for index in 0..lcp_array.len() {
         let lcp = lcp_array[index];
         if  lcp > max_len { max_len = lcp; }
-        let lcp = if lcp > u8::MAX as i32 { u8::MAX } else { lcp as u8 };
+        let lcp = if lcp > u16::MAX as i32 { u16::MAX } else { lcp as u16 };
 
         // Push new matches
         if lcp > stack.last().map_or(1, |m| m.len) {
